Removed excessive parameterization of methods
Having methods that mixed ReadablePartials/ReadableInstants was a bit
overkill.  First, it cluttered up the class, making it harder to tell
what method you wanted to use.  Second, its behavior is too random.
Depending on what partials you pass in, random fields get filled in
between it and ReadableInstant it usually needs to be converted into.

By keeping everything partial or instant, we have a grounding for
comparing the parameters and keep things a bit more succinct.