Removed "now" parameter from getRelativeTimeSpanString()
It didn't make sense that you'd need to pass in "now" when you
always want to compare the time to the current time.

By getting rid of it, it makes the API a lot easier to read/understand.

Also expanded test coverage to cover all API calls related to
getRelativeTimeSpanString()