Made tests much more consistent
Basically, the left side (what the assertion should equal) is mroe hardcoded.
This means that the tests have a consistent output they should be doing.

Most of this was achieved by setting a consistent "now" time for all tests
that doesn't vary.

The only place where this wasn't done is where the output can be platform
specific, e.g. formatDateRange() isn't fully implemented in our own DateUtils
and depends on Android.